Migrate to CommandLine::DispatchArg - Pass 1.
No attempt is made during this pass to fix any errors in the
program--or change any behaviour at all.  The only changes are
those required to bootstrap use of DispatchArg.

With few excepts, only "main" and "cmdline_*" functions are
modified.

Generally, command-line actions of the form:

 int cmdline_foo(int argc, char *argv[], bool simulate, ...)
 {
   ...
   if(failed)
     {
       fprintf(stderr, "E: Couldn't FOO\n");
       return -1;
     }

   _error->DumpErrors();

   return 0;
 }

become:

 bool cmdline_foo(CommandLine &cmdl)
 {
   const int argc = cmdl.FileSize();
   const bool simulate =
     aptcfg->FindB("Aptitude::Simulate", false);
   ...
   if(failed)
     return _error->Error("Couldn't FOO");

   return true;
 }

Process exit status is determined by the presence of pending
errors.  If there are pending errors then they are displayed
when cmdline_foo returns and the process exits with a status
of -1, otherwise, the status is 0.

Some actions return other status codes.  These have been
changed to the form:

 cmdline_foo(...)
 ...
   if(cond)
   {
     printf("No X\n");
     exit(1);
   }

Appropriate cleanup functions are registered with atexit(3)
to close the cache and display any errors.